requirements.txt
astroid==2.5.6
cffi==1.14.5
colorama==0.4.4
cryptography==3.4.7
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pika==1.2.0
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pylint==2.8.2
pymongo==3.11.4
pyOpenSSL==20.0.1
six==1.15.0
toml==0.10.2
typed-ast==1.4.3
wrapt==1.12.1
